+++ cli/src/commands/git/init.rs
@@ -61,9 +61,25 @@ pub struct GitInitArgs {
     /// the root of the `jj` repo along with the `.jj` directory. If the `.git`
     /// directory already exists, all the existing commits will be imported.
     ///
-    /// This option is mutually exclusive with `--git-repo`.
-    #[arg(long, conflicts_with = "git_repo")]
+    /// This option is mutually exclusive with `--git-repo` and `--no-colocate`.
+    #[arg(
+        long,
+        conflicts_with = "git_repo",
+        conflicts_with = "no_colocate",
+        default_value = "false"
+    )]
     colocate: bool,
+
+    /// Specifies that the `jj repo` should not be a valid `git` repo. The
+    /// `.git` directory will be placed underneath the `.jj` directory.
+    ///
+    /// This option is mutually exclusive with `--colocate` and `--git-repo`.
+    #[arg(
+        long = "no-colocate",
+        conflicts_with = "colocate",
+        default_value = "true"
+    )]
+    no_colocate: bool,
 
     /// Specifies a path to an **existing** git repository to be
     /// used as the backing git repo for the newly created `jj` repo.
@@ -95,13 +111,13 @@ pub fn cmd_git_init(
         .and_then(|_| wc_path.canonicalize())
         .map_err(|e| user_error_with_message("Failed to create workspace", e))?;
 
-    do_init(
-        ui,
-        command,
-        &wc_path,
-        args.colocate,
-        args.git_repo.as_deref(),
-    )?;
+    let colocate = if command.settings().git_settings().colocate {
+        !args.no_colocate
+    } else {
+        args.colocate
+    };
+
+    do_init(ui, command, &wc_path, colocate, args.git_repo.as_deref())?;
 
     let relative_wc_path = file_util::relative_path(cwd, &wc_path);
     writeln!(